backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Database dependencies
sqlalchemy==2.0.41
psycopg2-binary==2.9.9
alembic==1.16.3

# Testing dependencies
pytest-xdist==3.5.0
pytest-mock==3.12.0
pytest-cov==4.1.0
factory-boy==3.3.1
freezegun==1.2.2

# Image processing dependencies
Pillow==10.1.0
python-magic==0.4.27

# Google Cloud dependencies
google-cloud-storage==2.10.0
google-cloud-vision==3.4.5
google-cloud-aiplatform==1.38.1

# AI and ML dependencies
numpy==1.24.4